requirements.txt
google-api-python-client==2.125.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.4.0
